+++ build.ps1
@@ -316,8 +316,14 @@ process
             }
         }
 
-        # Loading Build Tasks defined in the .build/ folder (will override the ones imported above if same task name).
-        Get-ChildItem -Path '.build/' -Recurse -Include '*.ps1' -ErrorAction Ignore |
+        <#
+            Loading Build Tasks defined in the source/tasks folder (will override
+            the ones imported above if same task name).
+
+            Using 'source/tasks/' is changed specific for this module DscResource.DocGenerator
+            to be able to dogfood its own tasks
+        #>
+        Get-ChildItem -Path 'source/tasks/' -Recurse -Include '*.ps1' -ErrorAction Ignore |
             ForEach-Object {
                 "Importing file $($_.BaseName)" | Write-Verbose